Jumping Mechanic Added (Buggy) + some other animations
Added a few jump animations however the code implementation isn't correct so the character just keeps freefalling.
I've divided the jump animation into parts:
1 - Jump (up and forward) - there are two types of jumps. (i) Simple jump from standing idle, which triggers the jump up animation & (ii) Forward Jump from jogging or running. Note that (i) can be used while strafing and moving backwards as well.
2 (compulsory) - Free fall - all jumps transition into this
3 (optional) - Attack
4 - Land

This implementation will allow features like jump to land hard, jump to attack, jump to climb etc. to be implemented in the future but all that depends on whether this jumping mechanic is really enjoyable which we will see once the first digital prototype's implementation is complete.